feat: generalized macro rules for repeated patterns
Use of one macro rule with extra TT munching recursive macro.

At every invocation, it processes (munches) only a portion of the input,
generates intermediate output, and sends the remaining "unmunched" input
tail to another macro invocation. Eventually, the input is small enough to
not require any more macro invocations and reaches the base case at which
the recursion is terminated.

This macro rule should work for an arbitrary number of for $i in $iter
expressions as long as it stays within the macro recursion limit.